GH-288: Fix autoCommit Default Return Value
Fixes: https://github.com/spring-projects/spring-kafka/issues/288

The `DefaultKafkaConsumerFactory` incorrectly returns `false` if no `enable.auto.commit`
property is present. This causes the container to issue commits, which causes strange
offset behavior. Unread partitions are set to the end so messages can be lost.

Work-around is to explicitly add the property.

See: http://stackoverflow.com/questions/43405009/kafka-stop-consuming-message-from-new-assigned-patitions-after-rebalancing/43413899#43413899