GH-295: Unique client.id for Concurrent Container
Resolves: https://github.com/spring-projects/spring-kafka/issues/295

When JMX is enabled and a `client.id` property provided, we get an `InstanceAlreadyExistsException`
when using the `ConcurrentMessageListenerContainer` with `concurrency > 1`.

Add a mechanism to append the client index to the client.id.

Also fix private inner class ctors -> package to avoid synthetic ctor.

Fix deprecation warning and test

Conflicts:
	spring-kafka/src/main/java/org/springframework/kafka/core/DefaultKafkaConsumerFactory.java
	spring-kafka/src/main/java/org/springframework/kafka/listener/KafkaMessageListenerContainer.java
Resolved.