GH-380: Add timeout to record-getting test utils
* GH-380: Add timeout parameter to record-getting test utils

This parameter is useful when the test expects messages to be there but they
are not because the tested code is not correct. In this case the test will
run for a minute waiting for message before it fails. With a custom timeout,
a test can specify a shorter timeout to fail faster.

Closes GH-380.

* tweaking the javadoc to respect project conventions

See code review comment https://github.com/spring-projects/spring-kafka/pull/381#discussion_r130465766

*  add @since javadoc directives
*  add @author javadoc directive