GH-418: KafkaAdmin Improvements
Resolves https://github.com/spring-projects/spring-kafka/issues/418

Depending on the nature of the problem, `fatalIfBrokerNotAvailable` did not always work.

With a bad broker address, the client is not created.
With a good broker address, but the broker is down, the add fails.

Add more robust logic to the initialization code.

Also add the `autoCreate` flag, allowing users to suppress auto creation during
context initialization.