GH-701: Support ConsumerRecords<?, ?> (@Klistener)
Resolves https://github.com/spring-projects/spring-kafka/issues/701

Support passing the entire `ConsumerRecords<?, ?>` object to the listener.
Record filtering is not applied in this case.

Polishing - use the consumer records' map's keyset instead of creating new TopicPartitions when finding offsets to commit.

Optimization; the highest offset is always the last in the list; similar optimization for SeekToCurrentBatchErrorHandler.

Polishing - PR comments

* Polishing code style