Document (de)serialization config properties
When reading the reference manual, it is likely that one would encounter the
"Introduction" and start with the "consumer properties map" approach to
configuring deserialization. However, in the course of troubleshooting consumer
failures (eg. deserialization errors), one might read further into the manual
and encounter the "Serialization, Deserialization, and Message Conversion"
section, and pass instances of deserializers to the consumer factory in an
attempt to fix the failures, and might miss out that in fact consumer
properties map now has no effect.

Mention this in the documentation, and add a test case illustrating this is the
expected behavior.